Add minimal karma configuration to run an arbitrary JS file
Add PhantomJS as browser option for karma.(Automatically adds a dev dependency for karma-phantomjs-launcher)
Add a dummy JS file to file list of karma
PhantomJS native component should be installed for 'grunt karma' to work after this commit

Notice that, Frameworks: is configured as an empty [] in karma.conf.js
This leads to the following error as karma cannot find an appropriate test adaptor.

"You need to include some adapter that implements __karma__.start method!"

However karma launches PhantomJS and loads  spec/hello.js and executes it.

The karma:unit task fails here, but is a step forward from the previous state.